--- NOVR/Patches.cs
@@ -1,17 +1,0 @@
-using System.Reflection;
-using HarmonyLib;
-using UnityEngine;
-
-namespace NOVR;
-
-[HarmonyPatch(typeof(Camera), "set_fieldOfView")]
-public static class CameraPatches
-{
-    [HarmonyPrefix]
-    // Unity already prevents this, but it also nags you constantly about it.
-    // Some games try to change the FOV every frame, and all those logs can reduce performance.
-    private static bool PreventChangingFov()
-    {
-        return false;
-    }
-}
