+++ NOVR/Patches/HUD/HUDBombingStatePatch.cs
@@ -1,39 +1,40 @@
-using HarmonyLib;
 using System.Reflection;
+using HarmonyLib;
+using NOVR.PatchHelper;
+using NOVR.VrUi.HarmonyPatches;
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace NOVR.VrUi.HarmonyPatches;
+namespace NOVR.Patches.HUD;
 
-internal static class HUDBombingStateViewPositionPatch
+internal static class HUDBombingStatePatch
 {
-    private static readonly FieldInfo AlignmentBarField = AccessTools.Field(typeof(global::HUDBombingState), "alignmentBar");
-    private static readonly FieldInfo CcipPipperField = AccessTools.Field(typeof(global::HUDBombingState), "ccipPipper");
-    private static readonly FieldInfo CcipLineField = AccessTools.Field(typeof(global::HUDBombingState), "ccipLine");
-    private static readonly FieldInfo CcipFallTimeField = AccessTools.Field(typeof(global::HUDBombingState), "ccipFallTime");
-    private static readonly FieldInfo CcrpFallTimeField = AccessTools.Field(typeof(global::HUDBombingState), "ccrpFallTime");
-    private static readonly FieldInfo DropCountdownField = AccessTools.Field(typeof(global::HUDBombingState), "dropCountdown");
-    private static readonly FieldInfo CcrpCircleField = AccessTools.Field(typeof(global::HUDBombingState), "ccrpCircle");
-    private static readonly FieldInfo AverageTargetPositionField = AccessTools.Field(typeof(global::HUDBombingState), "averageTargetPosition");
-    private static readonly FieldInfo CcipImpactPointSmoothedField = AccessTools.Field(typeof(global::HUDBombingState), "ccipImpactPointSmoothed");
+    private static readonly FieldInfo AlignmentBarField = AccessTools.Field(typeof(HUDBombingState), "alignmentBar");
+    private static readonly FieldInfo CcipPipperField = AccessTools.Field(typeof(HUDBombingState), "ccipPipper");
+    private static readonly FieldInfo CcipLineField = AccessTools.Field(typeof(HUDBombingState), "ccipLine");
+    private static readonly FieldInfo CcipFallTimeField = AccessTools.Field(typeof(HUDBombingState), "ccipFallTime");
+    private static readonly FieldInfo CcrpFallTimeField = AccessTools.Field(typeof(HUDBombingState), "ccrpFallTime");
+    private static readonly FieldInfo DropCountdownField = AccessTools.Field(typeof(HUDBombingState), "dropCountdown");
+    private static readonly FieldInfo CcrpCircleField = AccessTools.Field(typeof(HUDBombingState), "ccrpCircle");
+    private static readonly FieldInfo AverageTargetPositionField = AccessTools.Field(typeof(HUDBombingState), "averageTargetPosition");
+    private static readonly FieldInfo CcipImpactPointSmoothedField = AccessTools.Field(typeof(HUDBombingState), "ccipImpactPointSmoothed");
 
-    [HarmonyPatch(typeof(global::HUDBombingState), nameof(global::HUDBombingState.UpdateWeaponDisplay))]
-    private static class UpdateWeaponDisplayPatch
+    
+    
+    
+    [PatchPostfix(typeof(HUDBombingState), nameof(HUDBombingState.UpdateWeaponDisplay))]
+    private static void UpdateWeaponDisplay(HUDBombingState __instance, Aircraft aircraft)
     {
-        [HarmonyPostfix]
-        private static void Postfix(global::HUDBombingState __instance, Aircraft aircraft)
-        {
-            var mainCamera = APIBus.MainCamera;
-            var cockpitHudCamera = APIBus.CockpitHudCamera;
-            if (mainCamera == null || cockpitHudCamera == null || aircraft == null)
-                return;
+        var mainCamera = APIBus.MainCamera;
+        var cockpitHudCamera = APIBus.CockpitHudCamera;
+        if (mainCamera == null || cockpitHudCamera == null || aircraft == null)
+            return;
 
-            __instance.transform.localPosition = Vector3.zero;
-            __instance.transform.rotation = cockpitHudCamera.transform.rotation;
+        __instance.transform.localPosition = Vector3.zero;
+        __instance.transform.rotation = cockpitHudCamera.transform.rotation;
 
-            UpdateCcrpDisplay(__instance, aircraft);
-            UpdateCcipDisplay(__instance);
-        }
+        UpdateCcrpDisplay(__instance, aircraft);
+        UpdateCcipDisplay(__instance);
     }
 
     private static void UpdateCcrpDisplay(global::HUDBombingState state, Aircraft aircraft)
@@ -53,15 +54,15 @@ internal static class HUDBombingStateViewPositionPatch
         var upperWorldPosition = averageTargetPosition.ToLocalPosition() + Vector3.up * verticalOffset;
         var lowerWorldPosition = averageTargetPosition.ToLocalPosition() + Vector3.up * verticalOffset * 0.9f;
 
-        if (!VrHudProjection.TryProjectToCockpitHud(upperWorldPosition, out var upperHudPosition) ||
-            !VrHudProjection.TryProjectToCockpitHud(lowerWorldPosition, out var lowerHudPosition))
+        if (!VrHudProjectionHelper.TryProjectToCockpitHud(upperWorldPosition, out var upperHudPosition) ||
+            !VrHudProjectionHelper.TryProjectToCockpitHud(lowerWorldPosition, out var lowerHudPosition))
         {
             alignmentBar.gameObject.SetActive(false);
             return;
         }
 
         alignmentBar.transform.position = upperHudPosition;
-        alignmentBar.transform.rotation = VrHudProjection.GetRotationAlongHudSegment(upperHudPosition, lowerHudPosition, cockpitHudCamera);
+        alignmentBar.transform.rotation = VrHudProjectionHelper.GetRotationAlongHudSegment(upperHudPosition, lowerHudPosition, cockpitHudCamera);
 
         if (ccrpCircle != null)
             ccrpCircle.transform.rotation = cockpitHudCamera.transform.rotation;
@@ -85,7 +86,7 @@ internal static class HUDBombingStateViewPositionPatch
 
         var ccipImpactPointSmoothed = (Vector3)CcipImpactPointSmoothedField.GetValue(state);
         var impactWorldPosition = ccipImpactPointSmoothed + Datum.origin.position;
-        if (!VrHudProjection.TryProjectToCockpitHud(impactWorldPosition, out var pipperHudPosition))
+        if (!VrHudProjectionHelper.TryProjectToCockpitHud(impactWorldPosition, out var pipperHudPosition))
         {
             ccipPipper.enabled = false;
             ccipLine.enabled = false;
@@ -104,8 +105,8 @@ internal static class HUDBombingStateViewPositionPatch
         }
 
         var normalizedLineDirection = lineDirection.normalized;
-        var lineStart = pipperHudPosition + normalizedLineDirection * VrHudProjection.ReferencePixelsToHudDistance(22.0f);
-        var lineEnd = velocityHudPosition - normalizedLineDirection * VrHudProjection.ReferencePixelsToHudDistance(8.0f);
+        var lineStart = pipperHudPosition + normalizedLineDirection * VrHudProjectionHelper.ReferencePixelsToHudDistance(22.0f);
+        var lineEnd = velocityHudPosition - normalizedLineDirection * VrHudProjectionHelper.ReferencePixelsToHudDistance(8.0f);
         var lineVector = lineEnd - lineStart;
         if (Vector3.Dot(lineDirection, lineVector) < 0.0f)
         {
@@ -113,7 +114,7 @@ internal static class HUDBombingStateViewPositionPatch
             return;
         }
 
-        VrHudProjection.SetVerticalLine(ccipLine.transform, lineStart, lineEnd, cockpitHudCamera);
+        VrHudProjectionHelper.SetVerticalLine(ccipLine.transform, lineStart, lineEnd, cockpitHudCamera);
 
         if (ccipFallTime != null)
             ccipFallTime.transform.rotation = cockpitHudCamera.transform.rotation;
