+++ NOVR/Patches/HUD/Map/DynamicMapRadarVisualizationRotationSafePatch.cs
@@ -1,10 +1,10 @@
-using HarmonyLib;
 using System.Collections;
 using System.Reflection;
+using HarmonyLib;
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace NOVR.VrUi.HarmonyPatches;
+namespace NOVR.Patches.HUD.Map;
 
 internal static class DynamicMapRadarVisualizationRotationSafePatch
 {
