+++ NOVR/VrCamera/TurretVrCameraPatch.cs
@@ -5,11 +5,11 @@ namespace NOVR.VrCamera;
 
 internal static class TurretVrCameraPatch
 {
-    [HarmonyPatch(typeof(global::Turret), "FixedUpdate")]
+    [HarmonyPatch(typeof(Turret), "FixedUpdate")]
     private static class FixedUpdatePatch
     {
         [HarmonyPrefix]
-        private static bool Prefix(global::Turret __instance)
+        private static bool Prefix(Turret __instance)
         {
             if (__instance == null)
             {
@@ -18,17 +18,17 @@ internal static class TurretVrCameraPatch
 
             var turret = Traverse.Create(__instance);
             if (!turret.Field("manual").GetValue<bool>() ||
-                turret.Field("target").GetValue<global::Unit>() != null)
+                turret.Field("target").GetValue<Unit>() != null)
             {
                 return true;
             }
 
-            var aircraft = turret.Field("aircraft").GetValue<global::Aircraft>();
-            var attachedUnit = turret.Field("attachedUnit").GetValue<global::Unit>();
+            var aircraft = turret.Field("aircraft").GetValue<Aircraft>();
+            var attachedUnit = turret.Field("attachedUnit").GetValue<Unit>();
             if (aircraft == null ||
                 attachedUnit == null ||
                 !aircraft.LocalSim ||
-                global::SceneSingleton<global::CameraStateManager>.i.currentState != global::SceneSingleton<global::CameraStateManager>.i.cockpitState)
+                SceneSingleton<CameraStateManager>.i.currentState != SceneSingleton<CameraStateManager>.i.cockpitState)
             {
                 return true;
             }
@@ -44,7 +44,7 @@ internal static class TurretVrCameraPatch
             var lastVectorSent = turret.Field("lastVectorSent").GetValue<float>();
             if (Time.timeSinceLevelLoad - lastVectorSent > 0.20000000298023224)
             {
-                var currentWeaponStation = turret.Field("currentWeaponStation").GetValue<global::WeaponStation>();
+                var currentWeaponStation = turret.Field("currentWeaponStation").GetValue<WeaponStation>();
                 var manualVector = turret.Field("manualVector").GetValue<Vector3>();
                 aircraft.SetTurretVector(currentWeaponStation.Number, manualVector);
                 turret.Field("lastVectorSent").SetValue(Time.timeSinceLevelLoad);
