+++ NOVR/VrCamera/VrCameraOffset.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using UnityEngine;
+﻿using UnityEngine;
 
 namespace NOVR.VrCamera;
