+++ NOVR/VrUi/Components/NOVRMainMenuBehavior.cs
@@ -1,6 +1,4 @@
-using System;
 using UnityEngine;
-using UnityEngine.UI;
 
 namespace NOVR.VrUi.SpecialBehavior;
