+++ NOVR/VrUi/Components/NOVRStatusDisplayBehavior.cs
@@ -1,4 +1,3 @@
-using System;
 using UnityEngine;
 
 namespace NOVR.VrUi.SpecialBehavior;
